factor_out
